Refactor discover endpoints into dedicated Django app
Move all feed discovery, search, and add site related endpoints from
rss_feeds app into new apps/discover/ app for better organization.

Changes:
- Create apps/discover/ with views.py, urls.py, tests.py
- Move 11 view functions: feed_autocomplete, search_feed, trending_sites,
  youtube_search, reddit_search, reddit_popular, podcast_search,
  newsletter_convert, google_news_feed, discover_feeds, discover_stories
- Update all frontend JS files to use new /discover/ endpoints
- Add HTTP 308 redirects for pre-existing URLs to maintain backward
  compatibility (preserves POST method)

New URL structure under /discover/:
- /autocomplete, /search_feed, /trending
- /youtube/search, /reddit/search, /reddit/popular
- /podcast/search, /newsletter/convert, /google-news/feed
- /similar/<feed_id>, /similar/feeds, /similar/stories/<hash>
